NeighborhoodExplorer have been deeply reviewed to simplify their implementation. Currently also relatedness can be specified through a lambda (in case of union). Still they are experimental.

+++ include/easylocal/helpers/kicker.hh
@@ -419,7 +419,7 @@ namespace EasyLocal {
       Kicker(NeighborhoodExplorer<Input, State, Move, CFtype>& ne, const MoveRelatedness& RelatedMoves = AllMovesRelated) : ne(ne), RelatedMoves(RelatedMoves) {}
       
       /** The modality of the @ref Move (warning: not the length of the @ref Move sequences) */
-      virtual unsigned int Modality() const
+      virtual size_t Modality() const
       {
         return ne.Modality();
       }
